Rename folder to plots
