+++ src/__test__/components/plots/ContinuousEmbeddingPlot.test.jsx
@@ -7,21 +7,34 @@ import thunk from 'redux-thunk';
 import { Provider } from 'react-redux';
 import { Vega } from 'react-vega';
 
-import CategoricalEmbeddingPlot from '../../../components/sharedPlots/CategoricalEmbeddingPlot';
+import ContinuousEmbeddingPlot from '../../../components/plots/ContinuousEmbeddingPlot';
 import { initialEmbeddingState } from '../../../redux/reducers/embeddings/initialState';
 import initialCellSetsState from '../../../redux/reducers/cellSets/initialState';
+import initialGeneExpressionState, { initialExpressionState } from '../../../redux/reducers/genes/initialState';
 import initialExperimentState from '../../../redux/reducers/experimentSettings/initialState';
 import { initialPlotConfigStates } from '../../../redux/reducers/componentConfig/initialState';
 
 jest.mock('localforage');
 const mockStore = configureStore([thunk]);
 
-describe('Categorical embedding', () => {
+describe('Continuous embedding plot', () => {
+  const config = initialPlotConfigStates.embeddingContinuous;
+
   const emptyStore = {
     cellSets: {
       ...initialCellSetsState,
     },
     embeddings: initialEmbeddingState,
+    genes: {
+      ...initialGeneExpressionState,
+      expression: {
+        loading: [],
+        error: false,
+        data: {
+          ...initialExpressionState,
+        },
+      },
+    },
     experimentSettings: {
       ...initialExperimentState,
     },
@@ -71,6 +84,24 @@ describe('Categorical embedding', () => {
         error: false,
       },
     },
+    genes: {
+      ...initialGeneExpressionState,
+      expression: {
+        loading: [],
+        error: false,
+        data: {
+          CST3: {
+            min: 1,
+            max: 6,
+            mean: 3.5,
+            stdev: 1.870828693387,
+            expression: [
+              1, 2, 3, 4, 5, 6,
+            ],
+          },
+        },
+      },
+    },
     experimentSettings: {
       ...initialExperimentState,
     },
@@ -78,14 +109,12 @@ describe('Categorical embedding', () => {
 
   configure({ adapter: new Adapter() });
 
-  const config = initialPlotConfigStates.embeddingCategorical;
-
   it('shows spinner when data is still loading', () => {
     const store = mockStore(emptyStore);
 
     const component = mount(
       <Provider store={store}>
-        <CategoricalEmbeddingPlot
+        <ContinuousEmbeddingPlot
           config={config}
         />
       </Provider>,
@@ -100,10 +129,15 @@ describe('Categorical embedding', () => {
   it('renders correctly when data is in the store', () => {
     const store = mockStore(mockedStore);
 
+    const chosenStore = {
+      ...config,
+      shownGene: 'CST3',
+    };
+
     const component = mount(
       <Provider store={store}>
-        <CategoricalEmbeddingPlot
-          config={config}
+        <ContinuousEmbeddingPlot
+          config={chosenStore}
         />
       </Provider>,
     );
