+++ src/components/plots/CategoricalEmbeddingPlot.jsx
