+++ src/components/plots/ContinuousEmbeddingPlot.jsx
